requirements.txt
asttokens==2.4.1
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
executing==2.0.1
filelock==3.15.4
Flask==3.0.3
fonttools==4.53.1
fsspec==2024.6.1
gunicorn==23.0.0
idna==3.7
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
matplotlib-inline==0.1.7
mkl==2021.4.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==26.1.0
requests==2.32.3
scipy==1.14.0
seaborn==0.13.2
six==1.16.0
stack-data==0.6.3
sympy==1.13.2
tbb==2021.13.1
torch==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.29
ultralytics-thop==2.0.0
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
